Added a plugin to allow enrollment of cohorts into courses using the API.
In order to use this feature, we must:

- Enable cohort sync: Site administration > Plugins > Enrolments > Manage enrol plugins
- Enable the following functions on the connectbox API
    - local_ws_enrolcohort_add_instance
    - local_ws_enrolcohort_delete_instance
    - local_ws_enrolcohort_get_instances

Note: in order for the 'Cohort sync' option to show up here, at least one Cohort must exist and
be available at either the site level or the category level in the category the course is in.
If there are no Cohorts available, then this option will not appear in the pull down list.